requirements.txt
flask
Flask-WTF
email_validator
requests
python-dotenv
